backend/lambda/requirements.txt
# Lambda Layer dependencies (production only)
psycopg2-binary==2.9.9
boto3==1.34.34
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
requests==2.31.0
email-validator==2.1.0
